Write first test
