--- tests/afg3021b.py
@@ -1,34 +1,0 @@
-import unittest
-
-from utils.channel_access import ChannelAccess
-from utils.ioc_launcher import get_default_ioc_dir
-from utils.test_modes import TestModes
-from utils.testing import get_running_lewis_and_ioc, skip_if_recsim
-
-
-DEVICE_PREFIX = "AFG3021B_01"
-
-
-IOCS = [
-    {
-        "name": DEVICE_PREFIX,
-        "directory": get_default_ioc_dir("AFG3021B"),
-        "macros": {},
-        "emulator": "Afg3021B",
-    },
-]
-
-
-TEST_MODES = [TestModes.RECSIM, TestModes.DEVSIM]
-
-
-class Afg3021BTests(unittest.TestCase):
-    """
-    Tests for the Afg3021B IOC.
-    """
-    def setUp(self):
-        self._lewis, self._ioc = get_running_lewis_and_ioc("Afg3021B", DEVICE_PREFIX)
-        self.ca = ChannelAccess(device_prefix=DEVICE_PREFIX)
-
-    def test_that_fails(self):
-        self.fail("You haven't implemented any tests!")
